Creates a file.

+++ 0x15-file_io/1-create_file.c
@@ -18,7 +18,7 @@ int create_file(const char *filename, char *text_content)
 {
 	int file_d;
 	size_t size = 0;
-	ssize_t length = 0;
+	int length = 0;
 
 	/** To get the size of text_content */
 	while (*(text_content + size) != '\0')
